Split daemon command modules
