+++ crates/roux-cli/src/daemon/status.rs
@@ -1,0 +1,148 @@
+use roux_runtime::host::RuntimeHost;
+
+use super::identity::DaemonIdentity;
+use super::protocol::Response;
+use super::unix_now_ms;
+
+pub(super) async fn handle_daemon_status(
+    host: &RuntimeHost,
+    identity: &DaemonIdentity,
+) -> Response {
+    let session_count = host.session_handle.list().await.map(|s| s.len()).unwrap_or(0);
+    let project_count = host.project_handle.list().await.map(|p| p.len()).unwrap_or(0);
+    let watch_count = host.watch_handle.list().await.map(|w| w.len()).unwrap_or(0);
+    let process_count = host.process_handle.list().await.map(|p| p.len()).unwrap_or(0);
+    let pty_count = host.pty_handle.list().await.map(|p| p.len()).unwrap_or(0);
+    let capabilities = vec![
+        "daemon-status",
+        "daemon-stop",
+        "session-list",
+        "session-poll",
+        "session-create",
+        "session-create-shell",
+        "session-reconnect-shell",
+        "session-panes-list",
+        "session-panes-create",
+        "session-archive",
+        "session-kill",
+        "session-restore",
+        "session-delete",
+        "session-worktree-exists",
+        "session-refresh-branch",
+        "session-rename",
+        "session-set-project",
+        "session-set-pinned-pr-url",
+        "session-set-smol-machine",
+        "alias-set",
+        "alias-unset",
+        "alias-claim",
+        "alias-list",
+        "alias-get",
+        "alias-whoami",
+        "alias-add-member",
+        "alias-remove-member",
+        "alias-mode",
+        "alias-events",
+        "mailbox-post",
+        "mailbox-peek",
+        "mailbox-read",
+        "mailbox-get",
+        "mailbox-read-state",
+        "mailbox-mark-read",
+        "mailbox-ack",
+        "mailbox-retract",
+        "mailbox-dismiss",
+        "mailbox-count",
+        "mailbox-clear",
+        "mailbox-reply",
+        "mailbox-sent",
+        "mailbox-events",
+        "bus-publish",
+        "bus-tail",
+        "bus-subscribe",
+        "bus-unsubscribe",
+        "bus-subscriptions",
+        "subscription-events",
+        "project-list",
+        "project-create",
+        "project-remove",
+        "project-rename",
+        "project-update",
+        "watch-list",
+        "watch-create",
+        "watch-find-or-create",
+        "watch-remove",
+        "watch-pause",
+        "watch-resume",
+        "watch-replace",
+        "watch-events",
+        "watch-remove-for-session",
+        "watch-cleanup-orphans",
+        "worktree-list",
+        "worktree-create",
+        "worktree-remove",
+        "worktree-list-branches",
+        "git-init",
+        "notes-read",
+        "notes-write",
+        "notes-append",
+        "notes-path",
+        "notes-search",
+        "notes-vault-root",
+        "hook-show",
+        "hook-preview",
+        "hook-run",
+        "hook-approve",
+        "hook-clear-approvals",
+        "hook-log-list",
+        "hook-log-read",
+        "run",
+        "shell",
+        "split",
+        "send",
+        "latest-output",
+        "daemon-process-start",
+        "daemon-process-output",
+        "daemon-process-list",
+        "daemon-process-kill",
+        "daemon-pty-spawn-shell",
+        "daemon-pty-spawn-task",
+        "daemon-pty-output",
+        "daemon-pty-attach",
+        "daemon-pty-list",
+        "daemon-pty-write",
+        "daemon-pty-resize",
+        "daemon-pty-detach",
+        "daemon-pty-attach-pane",
+        "daemon-pty-mark-read",
+        "daemon-pty-set-name",
+        "daemon-pty-kill",
+    ];
+    Response::success(serde_json::json!({
+        "kind": "roux-daemon",
+        "pid": std::process::id(),
+        "socket": identity.endpoint_display(),
+        "logPath": identity.log_path.to_string_lossy(),
+        "startedAtMs": identity.started_at_ms,
+        "uptimeMs": unix_now_ms().saturating_sub(identity.started_at_ms),
+        "sessionCount": session_count,
+        "projectCount": project_count,
+        "watchCount": watch_count,
+        "processCount": process_count,
+        "ptyCount": pty_count,
+        "capabilities": capabilities,
+    }))
+}
+
+pub(super) async fn handle_daemon_stop(identity: &DaemonIdentity) -> Response {
+    if identity.shutdown_tx.is_none() {
+        return Response::err("daemon shutdown channel unavailable");
+    }
+
+    Response::success(serde_json::json!({
+        "stopping": true,
+        "pid": std::process::id(),
+        "socket": identity.endpoint_display(),
+        "logPath": identity.log_path.to_string_lossy(),
+    }))
+}
